// src/app/(app)/client/(portal)/dashboard/types.ts
export interface KlaviyoData {
  storeRevenue: number
  storeOrders: number
  recoveryRate: number
  totalLeads: number
  engagedLeads: number
  engagementRate: number
  totalRevenue: number
  campaignRevenue: number
  flowRevenue: number
  smsRevenue: number
  emailsSent: number
  delivered: number
  opened: number
  clicked: number
  openRate: number
  clickRate: number
  clickToOpenRate: number
  conversionRate: number
  unsubscribeRate: number
  bounceRate: number
  bounces: number
  campaignsCount: number
  campaignDelivered: number
  campaignRevenuePercent: number
  flowsCount: number
  activeFlows: number
  flowDelivered: number
  flowRevenuePercent: number
  recentCampaigns: Array<{
    id: string
    name: string
    status: string
    sentAt: string
    recipients: number
    delivered: number
    opened: number
    clicked: number
    revenue: number
    openRate: number
    clickRate: number
  }>
  topFlows: Array<{
    id: string
    name: string
    revenue: number
    delivered: number
    openRate: number
    clickRate: number
  }>
}